🐛 fix : mail 서비스 의존성 구분
